refactor(config-ui): organize directory about gitlab component
