+++ config-ui/src/components/blueprints/create-workflow/DataScopes.jsx
@@ -30,7 +30,7 @@ import ConnectionTabs from '@/components/blueprints/ConnectionTabs'
 import BoardsSelector from '@/components/blueprints/BoardsSelector'
 import DataDomainsSelector from '@/components/blueprints/DataDomainsSelector'
 import NoData from '@/components/NoData'
-import { GitLabMillerColumns } from '@/components/miller-columns'
+import { GitLabMillerColumns } from '@/components/gitlab'
 import GitlabProject from '@/models/GitlabProject'
 import GitHubProject from '@/models/GithubProject'
 import JenkinsJobsSelector from '@/components/blueprints/JenkinsJobsSelector'
