+++ config-ui/src/components/gitlab/index.ts
@@ -16,4 +16,4 @@
  *
  */
 
-export * from './gitlab'
+export * from './miller-columns'
