+++ config-ui/src/components/miller-columns/index.ts
@@ -17,4 +17,6 @@
  */
 
 export * from './miller-columns'
-export * from './custom'
+export * from './components'
+export * from './hooks'
+export * from './types'
